+++ backend/src/recreation-resource/recreation-resource.service.ts
@@ -2,7 +2,7 @@ import { Injectable } from "@nestjs/common";
 import { Prisma } from "@prisma/client";
 import { PrismaService } from "src/prisma.service";
 import { RecreationResourceDto } from "./dto/recreation-resource.dto";
-import { PaginatedRecreationResourceDto } from "./dto/paginated-recreation-resouce.dto";
+import { PaginatedRecreationResourceDto } from "./dto/paginated-recreation-resource.dto";
 
 interface RecreationActivityWithDescription {
   with_description: {
